Start refactoring Scheduler
Extract ScheduledTask and common code. Remove timeoutTimer.
