+++ lib/scheduler/Scheduler.js
@@ -3,43 +3,20 @@
 /** @author John Hann */
 
 var base = require('@most/prelude');
+var ScheduledTask = require('./ScheduledTask');
+var runTask = require('../task').runTask;
 
 module.exports = Scheduler;
 
-function ScheduledTask(delay, period, task, scheduler) {
-	this.time = delay;
-	this.period = period;
-	this.task = task;
-	this.scheduler = scheduler;
-	this.active = true;
-}
-
-ScheduledTask.prototype.run = function() {
-	return this.task.run(this.time);
-};
-
-ScheduledTask.prototype.error = function(e) {
-	return this.task.error(this.time, e);
-};
-
-ScheduledTask.prototype.dispose = function() {
-	this.scheduler.cancel(this);
-	return this.task.dispose();
-};
-
-function runTask(task) {
-	try {
-		return task.run();
-	} catch(e) {
-		return task.error(e);
-	}
+function earliest(timeline) {
+	return timeline.length === 0 ? Infinity : timeline[0].time
 }
 
 function Scheduler(timer) {
 	this.timer = timer;
 
 	this._timer = null;
-	this._nextArrival = 0;
+	this._nextArrival = Infinity;
 	this._tasks = [];
 
 	var self = this;
