+++ lib/scheduler/Timer.js
@@ -2,9 +2,23 @@
 /** @author Brian Cavalier */
 /** @author John Hann */
 
-var defer = require('../defer');
+var defer = require('../task').defer;
 
 /*global setTimeout, clearTimeout*/
+
+module.exports = Timer;
+
+function Timer() {}
+
+Timer.prototype.now = Date.now;
+
+Timer.prototype.setTimer = function(f, dt) {
+	return dt <= 0 ? runAsTask(f) : setTimeout(f, dt);
+};
+
+Timer.prototype.clearTimer = function(t) {
+	return t instanceof Task ? t.cancel() : clearTimeout(t);
+};
 
 function Task(f) {
 	this.f = f;
@@ -12,11 +26,7 @@ function Task(f) {
 }
 
 Task.prototype.run = function() {
-	if(!this.active) {
-		return;
-	}
-	var f = this.f;
-	return f();
+	return this.active && this.f();
 };
 
 Task.prototype.error = function(e) {
@@ -32,13 +42,3 @@ function runAsTask(f) {
 	defer(task);
 	return task;
 }
-
-module.exports = {
-	now: Date.now,
-	setTimer: function(f, dt) {
-		return dt <= 0 ? runAsTask(f) : setTimeout(f, dt);
-	},
-	clearTimer: function(t) {
-		return t instanceof Task ? t.cancel() : clearTimeout(t);
-	}
-};
