+++ lib/scheduler/defaultScheduler.js
@@ -3,10 +3,6 @@
 /** @author John Hann */
 
 var Scheduler = require('./Scheduler');
-var setTimeoutTimer = require('./timeoutTimer');
-var nodeTimer = require('./nodeTimer');
+var Timer = require('./Timer');
 
-var isNode = typeof process === 'object'
-		&& typeof process.nextTick === 'function';
-
-module.exports = new Scheduler(isNode ? nodeTimer : setTimeoutTimer);
+module.exports = new Scheduler(new Timer());
