--- lib/scheduler/timeoutTimer.js
@@ -1,15 +1,0 @@
-/** @license MIT License (c) copyright 2010-2016 original author or authors */
-/** @author Brian Cavalier */
-/** @author John Hann */
-
-/*global setTimeout, clearTimeout*/
-
-module.exports = {
-	now: Date.now,
-	setTimer: function(f, dt) {
-		return setTimeout(f, dt);
-	},
-	clearTimer: function(t) {
-		return clearTimeout(t);
-	}
-};
