+++ lib/sink/DeferredSink.js
@@ -2,7 +2,7 @@
 /** @author Brian Cavalier */
 /** @author John Hann */
 
-var defer = require('../defer');
+var defer = require('../task').defer;
 
 module.exports = DeferredSink;
