+++ lib/task.js
@@ -2,7 +2,8 @@
 /** @author Brian Cavalier */
 /** @author John Hann */
 
-module.exports = defer;
+exports.defer = defer;
+exports.runTask = runTask;
 
 function defer(task) {
 	return Promise.resolve(task).then(runTask);
